basic_layout refactor
