+++ src/basic_layout/vertical_nav_bar.rs
@@ -1,0 +1,160 @@
+use crate::{PageId, Msg, Urls as RootUrls, Context};
+use seed::{prelude::*, *};
+use seed_styles::{pc, rem, em};
+use seed_styles::*;
+use crate::styles::{self, themes::{Color, Breakpoint}, global};
+use seed_hooks::{*, topo::nested as view};
+use std::rc::Rc;
+
+#[view]
+pub fn vertical_nav_bar(root_base_url: &Url, page_id: PageId) -> Node<Msg> {
+    nav![
+        C!["vertical-nav-bar", "vertical-nav-bar-container"],
+        s()
+            .bottom("0")
+            .left("0")
+            .position(CssPosition::Absolute)
+            .top(global::HORIZONTAL_NAV_BAR_SIZE)
+            .z_index("1")
+            .background_color(Color::BackgroundDark1)
+            .overflow_y(CssOverflowY::Auto)
+            .raw("scrollbar-width: none;")
+            .width(global::VERTICAL_NAV_BAR_SIZE),
+        vertical_nav_buttons(root_base_url, page_id),
+    ]
+}
+
+#[view]
+fn vertical_nav_buttons(root_base_url: &Url, page_id: PageId) -> Vec<Node<Msg>> {
+    vec![
+        { 
+            let active = page_id == PageId::Board;
+            vertical_nav_button(
+                "Board", 
+                active,
+                RootUrls::new(root_base_url).root(), 
+                true, 
+                vertical_nav_icon("ic_board", "0 0 1395 1024", active, vec![
+                    path![attrs!{At::D => "M1308.009 174.381l-1220.668 0c-48.237 0-87.341-39.104-87.341-87.341v0.301c0-48.237 39.104-87.341 87.341-87.341l1220.668-0c48.237 0 87.341 39.104 87.341 87.341v-0.301c0 48.237-39.104 87.341-87.341 87.341z"}],
+                    path![attrs!{At::D => "M936.358 599.341l-849.016 0c-48.237 0-87.341-39.104-87.341-87.341v0.301c0-48.237 39.104-87.341 87.341-87.341l849.016-0c48.237 0 87.341 39.104 87.341 87.341v-0.301c0 48.237-39.104 87.341-87.341 87.341z"}],
+                    path![attrs!{At::D => "M1308.009 1024h-1220.668c-48.237-0-87.341-39.104-87.341-87.341v0.301c0-48.237 39.104-87.341 87.341-87.341l1220.668 0c48.237 0 87.341 39.104 87.341 87.341v-0.301c0 48.237-39.104 87.341-87.341 87.341z"}],
+                ])
+            )
+        },
+        {
+            let active = page_id == PageId::Discover;
+            vertical_nav_button(
+                "Discover", 
+                active,
+                RootUrls::new(root_base_url).discover_urls().root(), 
+                false, 
+                vertical_nav_icon("ic_discover", "0 0 1025 1024", active, vec![
+                    path![attrs!{At::D => "M602.353 575.849c49.694-96.075 99.991-192.151 150.588-288.226 3.012-6.024 10.842-13.252 4.819-19.275s-13.553 0-19.275 4.216l-291.84 150.588c-10.241 5.534-18.27 14.048-23.055 24.371l-0.135 0.326q-64.753 124.386-129.506 248.471c-8.734 16.866-17.468 33.129-25.901 49.995-2.711 4.819-6.024 11.445 4.518 12.951 4.819-2.108 10.24-4.216 15.36-6.927l289.732-150.588c10.78-5.894 19.287-14.788 24.546-25.559l0.151-0.342z"}],
+                    path![attrs!{At::D => "M883.351 161.732c-90.543-95.747-216.891-156.82-357.52-161.708l-0.88-0.024c-3.070-0.066-6.687-0.104-10.314-0.104-138.198 0-263.562 54.947-355.438 144.186l0.123-0.119c-98.26 92.852-159.424 224.071-159.424 369.575 0 142.717 58.843 271.691 153.591 363.984l0.111 0.107c88.622 88.958 210.672 144.561 345.709 146.368l0.343 0.004h24.094c277.633-5.364 500.641-231.69 500.641-510.104 0-136.661-53.732-260.772-141.221-352.36l0.185 0.195zM242.748 783.059c-70.126-69.135-113.568-165.177-113.568-271.364 0-210.414 170.574-380.988 380.988-380.988 0.644 0 1.288 0.002 1.931 0.005l-0.099-0c210.913 0 381.892 170.979 381.892 381.892s-170.979 381.892-381.892 381.892v0 0c-0.446 0.002-0.975 0.003-1.503 0.003-104.66 0-199.368-42.605-267.728-111.418l-0.020-0.021z"}],
+                ])
+            )
+        },
+        {
+            let active = false;
+            vertical_nav_button(
+                "Library", 
+                active,
+                RootUrls::new(root_base_url).library(), 
+                false, 
+                vertical_nav_icon("ic_library", "0 0 1209 1024", active, vec![
+                    path![attrs!{At::D => "M1204.706 917.082l-190.645-826.729c-9.055-39.42-43.838-68.374-85.384-68.374-48.324 0-87.499 39.175-87.499 87.499 0 6.779 0.771 13.378 2.23 19.714l-0.114-0.589 191.548 827.633c11.135 36.317 44.369 62.266 83.664 62.266 48.237 0 87.341-39.104 87.341-87.341 0-4.971-0.415-9.846-1.213-14.591l0.071 0.513z"}],
+                    path![attrs!{At::D => "M674.334 0c-0-0-0-0-0.001-0-48.071 0-87.040 38.969-87.040 87.040 0 0.106 0 0.212 0.001 0.318l-0-0.016v849.318c-0.096 1.532-0.151 3.323-0.151 5.127 0 48.237 39.104 87.341 87.341 87.341s87.341-39.104 87.341-87.341c0-1.804-0.055-3.594-0.162-5.371l0.012 0.244v-849.318c0-48.237-39.104-87.341-87.341-87.341v0z"}],
+                    path![attrs!{At::D => "M87.944 0c-0.179-0.001-0.391-0.002-0.602-0.002-48.237 0-87.341 39.104-87.341 87.341 0 0.001 0 0.002 0 0.002l-0-0v849.318c-0.096 1.532-0.151 3.323-0.151 5.127 0 48.237 39.104 87.341 87.341 87.341s87.341-39.104 87.341-87.341c0-1.804-0.055-3.594-0.162-5.371l0.012 0.244v-849.318c0-0.090 0.001-0.197 0.001-0.303 0-47.859-38.627-86.697-86.406-87.038l-0.032-0z"}],
+                    path![attrs!{At::D => "M380.988 171.369c-48.002 0.171-86.869 39.038-87.040 87.024l-0 0.016v678.249c-0.096 1.532-0.151 3.323-0.151 5.127 0 48.237 39.104 87.341 87.341 87.341s87.341-39.104 87.341-87.341c0-1.804-0.055-3.594-0.162-5.371l0.012 0.244v-678.249c-0.171-48.108-39.209-87.040-87.341-87.040-0 0-0 0-0.001 0l0-0z"}],
+                ])
+            )
+        },
+        {
+            let active = false;
+            vertical_nav_button(
+                "Settings", 
+                active,
+                RootUrls::new(root_base_url).settings(),  
+                false, 
+                vertical_nav_icon("ic_settings", "0 0 1043 1024", active, vec![
+                    path![attrs!{At::D => "M791.492 901.421c-0.137 1.886-0.214 4.085-0.214 6.303 0 14.689 3.414 28.58 9.492 40.924l-0.242-0.544c1.442 2.027 2.306 4.553 2.306 7.281 0 5.548-3.572 10.262-8.542 11.967l-0.089 0.027c-37.735 21.585-81.411 40.158-127.33 53.451l-4.284 1.062c-2.114 1.002-4.593 1.587-7.209 1.587-7.903 0-14.559-5.341-16.556-12.61l-0.028-0.12c-20.88-43.535-64.606-73.060-115.229-73.060-26.819 0-51.703 8.287-72.23 22.44l0.428-0.279c-19.628 13.227-34.808 31.704-43.688 53.426l-0.284 0.786c-3.614 8.734-7.529 11.746-17.769 9.035-51.834-13.272-97.233-31.525-139.449-54.835l3.016 1.527c-14.758-7.831-8.734-16.866-5.12-26.805 4.846-12.398 7.654-26.752 7.654-41.762 0-32.050-12.804-61.11-33.576-82.344l0.021 0.021c-22.874-25.484-55.92-41.441-92.693-41.441-10.83 0-21.336 1.384-31.352 3.985l0.864-0.191h-5.722c-30.118 9.336-30.118 9.035-44.273-18.372-17.236-31.193-32.683-67.512-44.377-105.477l-1.101-4.152c-3.915-12.348-1.807-18.673 11.445-24.094 45.171-18.059 76.501-61.451 76.501-112.16 0-0.275-0.001-0.549-0.003-0.823l0 0.042c-0.157-51.84-32.003-96.203-77.176-114.748l-0.829-0.301c-13.553-4.819-15.962-10.842-12.047-23.793 13.962-48.504 31.914-90.674 54.24-130.036l-1.534 2.94c6.024-10.541 11.746-12.649 23.793-7.831 14.648 6.459 31.727 10.219 49.685 10.219 35.285 0 67.18-14.517 90.038-37.904l0.023-0.024c21.532-21.755 34.835-51.691 34.835-84.733 0-19.022-4.409-37.015-12.26-53.011l0.314 0.709c-4.216-9.638-3.012-15.059 6.024-20.48 39.702-23.013 85.609-42.536 133.977-56.195l4.263-1.029c13.252-3.614 14.758 5.12 18.372 13.252 16.261 41.325 53.282 71.221 97.87 77.036l0.614 0.065c6.241 1.121 13.425 1.762 20.759 1.762 40.852 0 77.059-19.886 99.469-50.507l0.242-0.347c7.452-9.232 13.404-20.047 17.264-31.809l0.204-0.718c3.012-8.433 8.132-9.939 16.264-8.132 52.584 13.65 98.681 32.83 141.232 57.456l-2.691-1.437c9.336 5.12 8.433 11.144 4.819 19.576-6.604 14.774-10.451 32.016-10.451 50.158 0 69.362 56.229 125.591 125.591 125.591 18.623 0 36.299-4.053 52.195-11.326l-0.784 0.321c10.24-4.518 15.962-3.012 21.384 6.927 22.212 37.657 40.917 81.17 53.87 127.095l0.944 3.916c2.711 10.24 0 15.36-10.24 19.878-46.208 16.823-78.61 60.371-78.61 111.487 0 0.299 0.001 0.599 0.003 0.898l-0-0.046c-0.106 1.871-0.166 4.060-0.166 6.264 0 49.766 30.792 92.34 74.362 109.71l0.797 0.28c12.951 6.024 16.264 11.746 12.047 25.6-14.446 47.781-32.562 89.199-54.858 127.907l1.55-2.918c-5.421 10.24-10.842 12.348-22.287 8.132-14.209-5.966-30.724-9.432-48.048-9.432-45.354 0-85.159 23.756-107.651 59.503l-0.31 0.527c-11.029 16.816-17.591 37.422-17.591 59.561 0 1.826 0.045 3.642 0.133 5.446l-0.010-0.254zM520.433 711.68c109.44-1.529 197.571-90.604 197.571-200.264 0-110.613-89.669-200.282-200.282-200.282s-200.282 89.669-200.282 200.282c0 0.205 0 0.411 0.001 0.616l-0-0.032c0.498 110.402 90.11 199.707 200.582 199.707 1.166 0 2.329-0.010 3.49-0.030l-0.175 0.002z"}],
+                ])
+            )
+        },
+        {
+            let active = page_id == PageId::Addons;
+            vertical_nav_button(
+                "Addons", 
+                active,
+                RootUrls::new(root_base_url).addons_urls().root(), 
+                false, 
+                vertical_nav_icon("ic_addons", "0 0 1043 1024", active, vec![
+                    path![attrs!{At::D => "M145.468 679.454c-40.056-39.454-80.715-78.908-120.471-118.664-33.431-33.129-33.129-60.235 0-90.353l132.216-129.807c5.693-5.938 12.009-11.201 18.865-15.709l0.411-0.253c23.492-15.059 41.864-7.529 48.188 18.974 0 7.228 2.711 14.758 3.614 22.287 3.801 47.788 37.399 86.785 82.050 98.612l0.773 0.174c10.296 3.123 22.128 4.92 34.381 4.92 36.485 0 69.247-15.94 91.702-41.236l0.11-0.126c24.858-21.654 40.48-53.361 40.48-88.718 0-13.746-2.361-26.941-6.701-39.201l0.254 0.822c-14.354-43.689-53.204-75.339-99.907-78.885l-0.385-0.023c-18.372-2.409-41.562 0-48.188-23.492s11.445-34.635 24.998-47.887q65.054-62.946 130.409-126.795c32.527-31.925 60.235-32.226 90.353 0 40.659 39.153 80.715 78.908 120.471 118.362 8.348 8.594 17.297 16.493 26.82 23.671l0.587 0.424c8.609 7.946 20.158 12.819 32.846 12.819 24.823 0 45.29-18.653 48.148-42.707l0.022-0.229c3.012-13.252 4.518-26.805 8.734-39.755 12.103-42.212 50.358-72.582 95.705-72.582 3.844 0 7.637 0.218 11.368 0.643l-0.456-0.042c54.982 6.832 98.119 49.867 105.048 104.211l0.062 0.598c0.139 1.948 0.218 4.221 0.218 6.512 0 45.084-30.574 83.026-72.118 94.226l-0.683 0.157c-12.348 3.915-25.299 5.722-37.948 8.433-45.779 9.638-60.235 46.984-30.118 82.824 15.265 17.569 30.806 33.587 47.177 48.718l0.409 0.373c31.925 31.925 64.452 62.946 96.075 94.871 13.698 9.715 22.53 25.511 22.53 43.369s-8.832 33.655-22.366 43.259l-0.164 0.111c-45.176 45.176-90.353 90.353-137.035 134.325-5.672 5.996-12.106 11.184-19.169 15.434l-0.408 0.227c-4.663 3.903-10.725 6.273-17.341 6.273-13.891 0-25.341-10.449-26.92-23.915l-0.012-0.127c-2.019-7.447-3.714-16.45-4.742-25.655l-0.077-0.848c-4.119-47.717-38.088-86.476-82.967-97.721l-0.76-0.161c-9.584-2.63-20.589-4.141-31.947-4.141-39.149 0-74.105 17.956-97.080 46.081l-0.178 0.225c-21.801 21.801-35.285 51.918-35.285 85.185 0 1.182 0.017 2.36 0.051 3.533l-0.004-0.172c1.534 53.671 40.587 97.786 91.776 107.115l0.685 0.104c12.649 2.409 25.901 3.313 38.249 6.626 22.588 6.325 30.118 21.685 18.372 41.864-4.976 8.015-10.653 14.937-17.116 21.035l-0.051 0.047c-44.875 44.574-90.353 90.353-135.228 133.12-10.241 14.067-26.653 23.106-45.176 23.106s-34.935-9.039-45.066-22.946l-0.111-0.159c-40.659-38.852-80.414-78.908-120.471-118.362z"}],
+                ])
+            )
+        },
+    ]
+}
+
+#[view]
+fn vertical_nav_button(title: &str, active: bool, url: Url, margin_top: bool, icon: Node<Msg>) -> Node<Msg> {
+    a![
+        C!["nav-tab-button", "nav-tab-button-container"],
+        attrs!{
+            At::TabIndex => -1,
+            At::Title => title,
+            At::Href => url,
+        },
+        IF!(margin_top => s().margin_top(rem(1))),
+        s()
+            .height(global::VERTICAL_NAV_BAR_SIZE)
+            .width(global::VERTICAL_NAV_BAR_SIZE)
+            .align_items(CssAlignItems::Center)
+            .background_color(if active { Color::SecondaryVariant1Dark5 } else { Color::BackgroundDark1 })
+            .display(CssDisplay::Flex)
+            .flex_direction(CssFlexDirection::Column)
+            .justify_content(CssJustifyContent::Center)
+            .cursor(CssCursor::Pointer),
+        s()
+            .hover()
+            .background_color(if active { Color::SecondaryVariant1Dark4 } else { Color::BackgroundLight2 }),
+        icon,
+        vertical_nav_label(title, active),
+    ]
+}
+
+#[view]
+fn vertical_nav_icon(icon: &str, view_box: &str, active: bool, paths: Vec<Node<Msg>>) -> Node<Msg> {
+    svg![
+        C!["icon"],
+        attrs!{
+            At::from("icon") => icon,
+            At::ViewBox => view_box,
+        },
+        s()
+            .overflow(CssOverflow::Visible)
+            .fill(if active { Color::SurfaceLight5_90 } else { Color::SecondaryLight5_90 })
+            .flex(CssFlex::None)
+            .height(rem(1.7))
+            .margin_bottom(rem(0.5))
+            .width(rem(1.7)),
+        paths,
+    ]
+}
+
+#[view]
+fn vertical_nav_label(title: &str, active: bool) -> Node<Msg> {
+    div![
+        C!["label"],
+        s()
+            .color(if active { Color::SurfaceLight5_90 } else { Color::SecondaryVariant1_90 })
+            .flex(CssFlex::None)
+            .font_size(rem(0.9))
+            .font_weight("500")
+            .letter_spacing(rem(0.01))
+            .max_height(em(2.4))
+            .padding("0 0.2rem")
+            .text_align(CssTextAlign::Center),
+        title,
+    ]
+}
